+++ api/src/main/java/org/allaymc/api/block/BlockPlaceHelper.java
@@ -1,7 +1,5 @@
 package org.allaymc.api.block;
 
-import com.google.common.collect.BiMap;
-import com.google.common.collect.HashBiMap;
 import lombok.experimental.UtilityClass;
 import org.allaymc.api.block.data.BlockFace;
 import org.allaymc.api.block.data.CompassRoseDirection;
@@ -19,15 +17,6 @@ import static org.allaymc.api.block.property.type.BlockPropertyTypes.*;
  */
 @UtilityClass
 public final class BlockPlaceHelper {
-    public static final BiMap<BlockFace, Integer> DIRECTION_4_MAPPER = HashBiMap.create(4);
-
-    static {
-        DIRECTION_4_MAPPER.put(BlockFace.EAST, 0);
-        DIRECTION_4_MAPPER.put(BlockFace.WEST, 1);
-        DIRECTION_4_MAPPER.put(BlockFace.SOUTH, 2);
-        DIRECTION_4_MAPPER.put(BlockFace.NORTH, 3);
-    }
-
     public static BlockState processFacingDirectionProperty(BlockState blockState, Vector3ic placeBlockPos, PlayerInteractInfo placementInfo) {
         if (placementInfo == null) {
             return blockState;
@@ -81,15 +70,6 @@ public final class BlockPlaceHelper {
                 MINECRAFT_CARDINAL_DIRECTION,
                 placementInfo.player().getHorizontalFace().opposite().toMinecraftCardinalDirection()
         );
-    }
-
-    public static BlockState processDirection4Property(BlockState blockState, Vector3ic placeBlockPos, PlayerInteractInfo placementInfo) {
-        if (placementInfo == null) {
-            return blockState;
-        }
-
-        var playerFace = placementInfo.player().getHorizontalFace();
-        return blockState.setPropertyValue(DIRECTION_4, DIRECTION_4_MAPPER.get(playerFace.opposite()));
     }
 
     public static BlockState processUpsideDownBitProperty(BlockState blockState, Vector3ic placeBlockPos, PlayerInteractInfo placementInfo) {